refactor: lazy-load vega plugins
