+++ frontend/package.json
@@ -99,7 +99,6 @@
     "lodash-es": "^4.17.21",
     "lucide-react": "^0.378.0",
     "lz-string": "^1.5.0",
-    "mathjs": "^12.4.2",
     "mermaid": "^10.9.0",
     "partysocket": "1.0.1",
     "plotly.js": "^2.32.0",
